Player1,2_screen変更
