+++ app/src/main/res/layout/dentaku_player1.xml
@@ -198,7 +198,7 @@
         android:layout_width="100dp"
         android:layout_height="50dp"
         android:text="1000"
-        android:layout_below="@+id/Player_cal2"
+        android:layout_below="@+id/Player_cal1"
         android:layout_toRightOf="@+id/Button_7"
         android:layout_toEndOf="@+id/Button_7" />
 
@@ -246,9 +246,11 @@
         android:layout_alignEnd="@+id/insert_point3" />
 
     <EditText
-        android:id="@+id/Player_cal2"
+        android:id="@+id/Player_cal1"
         android:layout_width="200dp"
         android:layout_height="100dp"
+        android:maxLength="4"
+        android:lines="1"
         android:layout_alignParentTop="true"
         android:layout_alignParentRight="true"
         android:layout_alignParentEnd="true" />
