+++ app/src/main/res/layout/life_setting.xml
@@ -58,7 +58,7 @@ android:layout_height="match_parent">
     android:layout_width="100dp"
     android:layout_height="50dp"
     android:text="9"
-    android:layout_below="@+id/Player_cal2"
+    android:layout_below="@+id/Player_cal1"
     android:layout_toLeftOf="@+id/Button_0"
     android:layout_toStartOf="@+id/Button_0" />
 
@@ -67,7 +67,7 @@ android:layout_height="match_parent">
     android:layout_width="100dp"
     android:layout_height="50dp"
     android:text="7"
-    android:layout_below="@+id/Player_cal2"
+    android:layout_below="@+id/Player_cal1"
     android:layout_toLeftOf="@+id/Button_8"
     android:layout_toStartOf="@+id/Button_8" />
 
@@ -115,13 +115,13 @@ android:layout_height="match_parent">
     android:layout_height="50dp"
     android:text="return_page"
     android:layout_alignParentTop="true"
-    android:layout_alignRight="@+id/Player_cal2"
-    android:layout_alignEnd="@+id/Player_cal2"
-    android:layout_alignLeft="@+id/Player_cal2"
-    android:layout_alignStart="@+id/Player_cal2" />
+    android:layout_alignRight="@+id/Player_cal1"
+    android:layout_alignEnd="@+id/Player_cal1"
+    android:layout_alignLeft="@+id/Player_cal1"
+    android:layout_alignStart="@+id/Player_cal1" />
 
 <Button
-    android:id="@+id/Player_cal2"
+    android:id="@+id/Player_cal1"
     android:layout_width="200dp"
     android:layout_height="100dp"
     android:text="Life_Point"
